backend/requirements.txt
# Typhoon OCR Backend Dependencies

fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
python-dotenv>=1.0.0
openai>=1.3.0
pypdf>=3.17.0
pillow>=10.1.0

# From packages/typhoon_ocr
ftfy>=6.1.0
